refactor CUB tests
