+++ tests/cupy_tests/core_tests/test_ndarray_reduction.py
@@ -1,5 +1,6 @@
 import unittest
 
+import cupy
 from cupy import testing
 
 
@@ -101,3 +102,27 @@ class TestArrayReduction(unittest.TestCase):
     def test_min_nan_imag(self, xp, dtype):
         a = xp.array([float('nan')*1.j, 1.j, -1.j], dtype)
         return a.min()
+
+
+# This class compares CUB results against NumPy's
+@unittest.skipIf(cupy.cuda.cub_enabled is False, 'The CUB module is not built')
+@testing.parameterize(*testing.product({
+    'shape': [(10,), (10, 20), (10, 20, 30), (10, 20, 30, 40)],
+}))
+@testing.gpu
+class TestCUBreduction(unittest.TestCase):
+    @testing.for_contiguous_axes()
+    @testing.for_dtypes('bhilBHILefdFD')
+    @testing.numpy_cupy_allclose(rtol=1E-5)
+    def test_cub_min(self, xp, dtype, axis):
+        assert cupy.cuda.cub_enabled
+        a = testing.shaped_random(self.shape, xp, dtype)
+        return a.min(axis=axis)
+
+    @testing.for_contiguous_axes()
+    @testing.for_dtypes('bhilBHILefdFD')
+    @testing.numpy_cupy_allclose(rtol=1E-5)
+    def test_cub_max(self, xp, dtype, axis):
+        assert cupy.cuda.cub_enabled
+        a = testing.shaped_random(self.shape, xp, dtype)
+        return a.max(axis=axis)
